Remove unused service
