+++ controller/organization/manager.py
@@ -1,5 +1,5 @@
 from typing import Any, List, Dict, Optional, Union
-from controller.misc import config_service
+from config_handler import get_config_value
 
 from submodules.model import enums
 from submodules.model.business_objects import organization, general, user
@@ -102,7 +102,7 @@ def get_overview_stats(org_id: str) -> List[Dict[str, Union[str, int]]]:
 
 
 def can_create_local(org: bool = True) -> bool:
-    if config_service.get_config_value("is_managed"):
+    if get_config_value("is_managed"):
         return False
     existing_orgs = organization.get_all()
     checkvalue = 0 if org else 1
